Add caching for data fetched from MongoDB
This adds caching for data from MongoDB. To do this efficiently, we do
not immediately decode the BSON we receive from MongoDB, since we may
want to have the binary data to store it in Redis.
We still need to serialize this data, since we are dealing with a
list of BSON objects. To do this, we use pickle.
We re-use the existing Redis connection that we already have.

If this code is run with different Python versions, they must use the
same pickle versions. Otherwise, they may not share the same cache.